requirements.txt
aiofiles==24.1.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.29.0 ; python_version >= "3.12" and python_version < "4.0"
html5tagger==1.3.0 ; python_version >= "3.12" and python_version < "4.0"
httptools>=0.6.1 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
sanic-ext==23.12.0 ; python_version >= "3.12" and python_version < "4.0"
sanic-routing==23.12.0 ; python_version >= "3.12" and python_version < "4.0"
sanic==24.6.0 ; python_version >= "3.12" and python_version < "4.0"
setuptools==74.0.0 ; python_version >= "3.12" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.12" and python_version < "4.0"
tracerite==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
ujson==5.10.0 ; sys_platform != "win32" and implementation_name == "cpython" and python_version >= "3.12" and python_version < "4.0"
uvloop==0.20.0 ; sys_platform != "win32" and implementation_name == "cpython" and python_version >= "3.12" and python_version < "4.0"
websockets==13.0 ; python_version >= "3.12" and python_version < "4.0"
